Always show the base exception instead of the outer exception when an error is raised
There are some cases where the error gets wrapped with another exception (For example if an object is being cast to a different type and raises an error in the process). This change will now always show the base error.

Also, add on additional information to the error message to help users identify what is causing the validation error.