Validate that a record can be serialized by Canary before passing the record to the controller
When an object is returned from a controller, it is automatically serialized. If serialization is not tested and verified ahead of time, then this will throw an exception outside of the scope of our code. By validating ahead of time and returning a nicer error to the user this behavior can be avoided.