Do not count DeathRecord comparison in message testing and instead rely on sub elements to reflect errors
Currently, if there is an error in the DeathRecord in a Message it will show up twice (once in the record and once on the specific element in the record that caused the error). This removes the count from the DeathRecord and just relies on the failure in the record count itself.